requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=0.19.0

# LLM/API integration
google-generativeai>=0.3.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-community>=0.0.13
langchain-core>=0.1.18

# Web API
fastapi>=0.104.1
uvicorn>=0.23.2

# Database and CLI are part of standard library
# sqlite3
# argparse

# Frontend
# Note: Frontend uses CDN-hosted libraries (Tailwind, Alpine.js)

# Added from the code block
requests>=2.26.0
python-dateutil>=2.8.2

# Added from the code block
flask>=2.0.0

# Added from the code block
pytest>=6.2.5
